changes for registering extension packages
