+++ .ci/ci.jl
@@ -1,11 +1,11 @@
 using Pkg
 
 if ARGS[1] == "full"
-    pkgs = ["MadNLPHSL","MadNLPPardiso","MadNLPMumps","MadNLPGPU","MadNLPGraphs","MadNLPIterative"]
+    pkgs = ["MadNLPHSL","MadNLPPardiso","MadNLPMumps","MadNLPGPU","MadNLPGraph","MadNLPKrylov"]
 elseif ARGS[1] == "basic"
-    pkgs = ["MadNLPMumps","MadNLPGraphs","MadNLPIterative"]
+    pkgs = ["MadNLPMumps","MadNLPGraph","MadNLPKrylov"]
 else
-    error("proper argument should be ginven - full or basic")
+    error("proper argument should be given - full or basic")
 end
 
 Pkg.develop(PackageSpec(path=joinpath(@__DIR__,"..")))
@@ -15,12 +15,3 @@ Pkg.build()
 
 Pkg.test.("MadNLP", coverage=true)
 Pkg.test.(pkgs, coverage=true)
-
-
-
-
-
-
-
-
-
