+++ .github/workflows/test.yml
@@ -28,7 +28,7 @@ jobs:
       - run: julia-${JULIA_VERSION} --color=yes --project=.ci .ci/ci.jl full
       - uses: julia-actions/julia-processcoverage@v1
         with:
-          directories: src,lib/MadNLPGPU/src,lib/MadNLPHSL/src,lib/MadNLPPardiso/src,lib/MadNLPMumps/src,lib/MadNLPIterative/src
+          directories: src,lib/MadNLPGPU/src,lib/MadNLPHSL/src,lib/MadNLPPardiso/src,lib/MadNLPMumps/src,lib/MadNLPKrylov/src,lib/MadNLPGraph
       - uses: codecov/codecov-action@v1
         with:
           file: lcov.info
