+++ codecov.yml
@@ -1,0 +1,2 @@
+ignore:
+  - "src/LinearSolvers/pardisomkl.jl"  # cannot test because the default BLAS vendor is OpenBLAS
