+++ lib/MadNLPGraph/Project.toml
@@ -1,4 +1,4 @@
-name = "MadNLPGraphs"
+name = "MadNLPGraph"
 uuid = "2f590f30-ea46-438a-a78d-e5bfe3109704"
 authors = ["Sungho Shin <sungho.shin.ss@gmail.com>"]
 version = "0.1.0"
@@ -16,6 +16,7 @@ JuMP = "~0.21"
 Plasmo = "~0.3, ~0.4"
 LightGraphs = "~1.3"
 Metis = "~1.0"
+julia = "1.3"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
