+++ lib/MadNLPGraph/src/graphtools.jl
