+++ lib/MadNLPHSL/Project.toml
@@ -11,6 +11,7 @@ Libdl = "8f399da3-3557-5675-b5ff-fb832c97cbdb"
 
 [compat]
 MadNLP = "~0.2"
+BinaryProvider = "0.5"
 julia = "1.3"
 
 [extras]
