+++ lib/MadNLPKrylov/Project.toml
@@ -1,4 +1,4 @@
-name = "MadNLPIterative"
+name = "MadNLPKrylov"
 uuid = "1888cb03-ce40-4a36-8d45-ae8231a0e17c"
 authors = ["Sungho Shin <sungho.shin.ss@gmail.com>"]
 version = "0.1.0"
