+++ lib/MadNLPPardiso/Project.toml
@@ -11,6 +11,7 @@ Libdl = "8f399da3-3557-5675-b5ff-fb832c97cbdb"
 
 [compat]
 MadNLP = "~0.2"
+BinaryProvider = "0.5"
 MadNLPTests = "~0.1"
 julia = "1.3"
