+++ src/Interfaces/interfaces.jl
@@ -2,6 +2,4 @@
 # Created by Sungho Shin (sungho.shin@wisc.edu)
 
 include("MOI_interface.jl")
-# include("Plasmo_interface.jl")
 include("NLPModels_interface.jl")
-# include("cpp_interface.jl") # will be supported once Cxx supports Julia >= 1.4
